app/backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
uvloop==0.16.0
alembic==1.7.3
httptools==0.1.0
psycopg2-binary==2.9.1
SQLAlchemy==1.4.23
passlib==1.7.4
bcrypt==3.2.0
python-jose==3.3.0
python-multipart==0.0.5
pandas==1.3.3
python-dotenv==0.19.2
fastapi-mail==1.0.2